# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company Ship API",
        "description": "API quản lý vận tải - công ty, nhân sự, xe, chuyến đi, lương",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Danh sách hóa đơn",
                "operationId": "2e947f49543378e8d1ac24fc3ebcdcc7",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm theo code",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trip_id",
                        "in": "query",
                        "description": "Lọc theo chuyến xe",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Lọc theo khách hàng",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sắp xếp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi/trang",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Tạo hóa đơn mới",
                "operationId": "15892ddc0ad0649e0919d15345232fc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "customer_id",
                                    "total_amount"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "INV001"
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "trip_id": {
                                        "type": "integer"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "example": 5000000
                                    },
                                    "tax_amount": {
                                        "type": "number"
                                    },
                                    "issued_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "sent",
                                            "paid",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            }
        },
        "/api/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Chi tiết hóa đơn",
                "operationId": "5d25d392f44d2cf516114d871e72ecb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Cập nhật hóa đơn",
                "operationId": "e5aa8abff3efb646c72e59724ecb13e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "trip_id": {
                                        "type": "integer"
                                    },
                                    "total_amount": {
                                        "type": "number"
                                    },
                                    "tax_amount": {
                                        "type": "number"
                                    },
                                    "issued_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Xóa hóa đơn",
                "operationId": "410b8a2e4b097016c3dea391349a789b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Dùng token trả về từ POST /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}